Remember previous alarm time
Since the timer unit is disabled by setting the time to a dummy entry, the
previous time is lost when the app starts again afterwards.
Save the alarm time also to a comment entry where it is preserved even if
the time is set to a dummy entry or a delayed time for snoozing. On start,
restore the alarm time from there if the alarm timer is disabled.